+++ src/main/java/mate/academy/model/Book.java
@@ -1,9 +1,12 @@
-package mate.academy.spring_boot_intro.model;
+package mate.academy.model;
 
-import jakarta.persistence.*;
-import lombok.Data;
-
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
 import java.math.BigDecimal;
+import lombok.Data;
 
 @Entity
 @Data
