+++ src/main/java/mate/academy/repository/BookRepository.java
@@ -1,0 +1,10 @@
+package mate.academy.repository;
+
+import java.util.List;
+import mate.academy.model.Book;
+
+public interface BookRepository {
+    Book save(Book book);
+
+    List<Book> findAll();
+}
