+++ src/main/java/mate/academy/repository/impl/BookDaoImpl.java
@@ -1,16 +1,14 @@
-package mate.academy.spring_boot_intro.repository.impl;
+package mate.academy.repository.impl;
 
 import jakarta.persistence.criteria.CriteriaQuery;
-import mate.academy.spring_boot_intro.model.Book;
-import mate.academy.spring_boot_intro.repository.BookRepository;
+import java.util.List;
+import mate.academy.model.Book;
+import mate.academy.repository.BookRepository;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
 import org.hibernate.Transaction;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
-
-import java.util.List;
-import java.util.Optional;
 
 @Repository
 public class BookDaoImpl implements BookRepository {
@@ -23,7 +21,7 @@ public class BookDaoImpl implements BookRepository {
     }
 
     @Override
-    public Book save (Book book) {
+    public Book save(Book book) {
         Session session = null;
         Transaction transaction = null;
         try {
