+++ src/main/java/mate/academy/service/BookService.java
@@ -1,0 +1,10 @@
+package mate.academy.service;
+
+import java.util.List;
+import mate.academy.model.Book;
+
+public interface BookService {
+    Book save(Book book);
+
+    List<Book> getAll();
+}
