+++ src/main/java/mate/academy/service/impl/BookServiceImpl.java
@@ -1,12 +1,11 @@
-package mate.academy.spring_boot_intro.service.impl;
-
-import mate.academy.spring_boot_intro.model.Book;
-import mate.academy.spring_boot_intro.repository.BookRepository;
-import mate.academy.spring_boot_intro.service.BookService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
+package mate.academy.service.impl;
 
 import java.util.List;
+import mate.academy.model.Book;
+import mate.academy.repository.BookRepository;
+import mate.academy.service.BookService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
 @Service
 public class BookServiceImpl implements BookService {
