--- src/main/java/mate/academy/spring_boot_intro/repository/BookRepository.java
@@ -1,12 +1,0 @@
-package mate.academy.spring_boot_intro.repository;
-
-import mate.academy.spring_boot_intro.model.Book;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface BookRepository {
-    Book save(Book book);
-
-    List<Book> findAll();
-}
