--- src/main/java/mate/academy/spring_boot_intro/service/BookService.java
@@ -1,12 +1,0 @@
-package mate.academy.spring_boot_intro.service;
-
-import mate.academy.spring_boot_intro.model.Book;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface BookService {
-    Book save(Book book);
-
-    List<Book> getAll();
-}
